Create README.md
This project is made very simply and conveniently. C++ has been used in this. In this, with the help of a switch case, we repeatedly take input from the user, and according to that input minus case, we run the functions. Function and switch cases have been used in this project so that the program can be easily understood and read. This program is run only once, meaning that no file is saved in it, and does not have a file handling system. This is a basic-level project